fix: tests work on 2022.2+
