+++ Assets/Mirror/Tests/Editor/Attributes/AttributeTests_MonoBehaviour.cs
@@ -1,162 +1,13 @@
 // Tests for [Server], [Client], [ServerCallback], [ClientCallback] attributes.
 // Generated by AttributeTestGenerator.cs
 
+using Mirror.Tests.EditorBehaviours.Attributes;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
 
-namespace Mirror.Tests.Attributes
+namespace Mirror.Tests.EditorBehaviours.Attributes
 {
-    public class AttributeBehaviour_MonoBehaviour : MonoBehaviour
-    {
-        public static readonly float Expected_float = 2020f;
-        public static readonly ClassWithNoConstructor Expected_ClassWithNoConstructor = new ClassWithNoConstructor { a = 10 };
-        public static readonly ClassWithConstructor Expected_ClassWithConstructor = new ClassWithConstructor(29);
-
-        [Client]
-        public float Client_float_Function()
-        {
-            return Expected_float;
-        }
-
-        [Client]
-        public void Client_float_out_Function(out float value)
-        {
-            value = Expected_float;
-        }
-
-        [Client]
-        public ClassWithNoConstructor Client_ClassWithNoConstructor_Function()
-        {
-            return Expected_ClassWithNoConstructor;
-        }
-
-        [Client]
-        public void Client_ClassWithNoConstructor_out_Function(out ClassWithNoConstructor value)
-        {
-            value = Expected_ClassWithNoConstructor;
-        }
-
-        [Client]
-        public ClassWithConstructor Client_ClassWithConstructor_Function()
-        {
-            return Expected_ClassWithConstructor;
-        }
-
-        [Client]
-        public void Client_ClassWithConstructor_out_Function(out ClassWithConstructor value)
-        {
-            value = Expected_ClassWithConstructor;
-        }
-
-        [Server]
-        public float Server_float_Function()
-        {
-            return Expected_float;
-        }
-
-        [Server]
-        public void Server_float_out_Function(out float value)
-        {
-            value = Expected_float;
-        }
-
-        [Server]
-        public ClassWithNoConstructor Server_ClassWithNoConstructor_Function()
-        {
-            return Expected_ClassWithNoConstructor;
-        }
-
-        [Server]
-        public void Server_ClassWithNoConstructor_out_Function(out ClassWithNoConstructor value)
-        {
-            value = Expected_ClassWithNoConstructor;
-        }
-
-        [Server]
-        public ClassWithConstructor Server_ClassWithConstructor_Function()
-        {
-            return Expected_ClassWithConstructor;
-        }
-
-        [Server]
-        public void Server_ClassWithConstructor_out_Function(out ClassWithConstructor value)
-        {
-            value = Expected_ClassWithConstructor;
-        }
-
-        [ClientCallback]
-        public float ClientCallback_float_Function()
-        {
-            return Expected_float;
-        }
-
-        [ClientCallback]
-        public void ClientCallback_float_out_Function(out float value)
-        {
-            value = Expected_float;
-        }
-
-        [ClientCallback]
-        public ClassWithNoConstructor ClientCallback_ClassWithNoConstructor_Function()
-        {
-            return Expected_ClassWithNoConstructor;
-        }
-
-        [ClientCallback]
-        public void ClientCallback_ClassWithNoConstructor_out_Function(out ClassWithNoConstructor value)
-        {
-            value = Expected_ClassWithNoConstructor;
-        }
-
-        [ClientCallback]
-        public ClassWithConstructor ClientCallback_ClassWithConstructor_Function()
-        {
-            return Expected_ClassWithConstructor;
-        }
-
-        [ClientCallback]
-        public void ClientCallback_ClassWithConstructor_out_Function(out ClassWithConstructor value)
-        {
-            value = Expected_ClassWithConstructor;
-        }
-
-        [ServerCallback]
-        public float ServerCallback_float_Function()
-        {
-            return Expected_float;
-        }
-
-        [ServerCallback]
-        public void ServerCallback_float_out_Function(out float value)
-        {
-            value = Expected_float;
-        }
-
-        [ServerCallback]
-        public ClassWithNoConstructor ServerCallback_ClassWithNoConstructor_Function()
-        {
-            return Expected_ClassWithNoConstructor;
-        }
-
-        [ServerCallback]
-        public void ServerCallback_ClassWithNoConstructor_out_Function(out ClassWithNoConstructor value)
-        {
-            value = Expected_ClassWithNoConstructor;
-        }
-
-        [ServerCallback]
-        public ClassWithConstructor ServerCallback_ClassWithConstructor_Function()
-        {
-            return Expected_ClassWithConstructor;
-        }
-
-        [ServerCallback]
-        public void ServerCallback_ClassWithConstructor_out_Function(out ClassWithConstructor value)
-        {
-            value = Expected_ClassWithConstructor;
-        }
-    }
 
     public class AttributeTest_MonoBehaviour
     {
@@ -188,7 +39,7 @@ namespace Mirror.Tests.Attributes
             float expected = active ? AttributeBehaviour_MonoBehaviour.Expected_float : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Single Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Client_float_Function()' called when client was not active");
+                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Single Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Client_float_Function()' called when client was not active");
             }
             float actual = behaviour.Client_float_Function();
             Assert.AreEqual(expected, actual);
@@ -203,7 +54,7 @@ namespace Mirror.Tests.Attributes
             float expected = active ? AttributeBehaviour_MonoBehaviour.Expected_float : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Void Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Client_float_out_Function(System.Single&)' called when client was not active");
+                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Void Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Client_float_out_Function(System.Single&)' called when client was not active");
             }
             behaviour.Client_float_out_Function(out float actual);
             Assert.AreEqual(expected, actual);
@@ -218,7 +69,7 @@ namespace Mirror.Tests.Attributes
             ClassWithNoConstructor expected = active ? AttributeBehaviour_MonoBehaviour.Expected_ClassWithNoConstructor : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Client] function 'Mirror.Tests.Attributes.ClassWithNoConstructor Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Client_ClassWithNoConstructor_Function()' called when client was not active");
+                LogAssert.Expect(LogType.Warning, "[Client] function 'Mirror.Tests.EditorBehaviours.Attributes.ClassWithNoConstructor Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Client_ClassWithNoConstructor_Function()' called when client was not active");
             }
             ClassWithNoConstructor actual = behaviour.Client_ClassWithNoConstructor_Function();
             Assert.AreEqual(expected, actual);
@@ -233,7 +84,7 @@ namespace Mirror.Tests.Attributes
             ClassWithNoConstructor expected = active ? AttributeBehaviour_MonoBehaviour.Expected_ClassWithNoConstructor : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Void Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Client_ClassWithNoConstructor_out_Function(Mirror.Tests.Attributes.ClassWithNoConstructor&)' called when client was not active");
+                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Void Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Client_ClassWithNoConstructor_out_Function(Mirror.Tests.EditorBehaviours.Attributes.ClassWithNoConstructor&)' called when client was not active");
             }
             behaviour.Client_ClassWithNoConstructor_out_Function(out ClassWithNoConstructor actual);
             Assert.AreEqual(expected, actual);
@@ -248,7 +99,7 @@ namespace Mirror.Tests.Attributes
             ClassWithConstructor expected = active ? AttributeBehaviour_MonoBehaviour.Expected_ClassWithConstructor : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Client] function 'Mirror.Tests.Attributes.ClassWithConstructor Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Client_ClassWithConstructor_Function()' called when client was not active");
+                LogAssert.Expect(LogType.Warning, "[Client] function 'Mirror.Tests.EditorBehaviours.Attributes.ClassWithConstructor Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Client_ClassWithConstructor_Function()' called when client was not active");
             }
             ClassWithConstructor actual = behaviour.Client_ClassWithConstructor_Function();
             Assert.AreEqual(expected, actual);
@@ -263,7 +114,7 @@ namespace Mirror.Tests.Attributes
             ClassWithConstructor expected = active ? AttributeBehaviour_MonoBehaviour.Expected_ClassWithConstructor : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Void Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Client_ClassWithConstructor_out_Function(Mirror.Tests.Attributes.ClassWithConstructor&)' called when client was not active");
+                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Void Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Client_ClassWithConstructor_out_Function(Mirror.Tests.EditorBehaviours.Attributes.ClassWithConstructor&)' called when client was not active");
             }
             behaviour.Client_ClassWithConstructor_out_Function(out ClassWithConstructor actual);
             Assert.AreEqual(expected, actual);
@@ -278,7 +129,7 @@ namespace Mirror.Tests.Attributes
             float expected = active ? AttributeBehaviour_MonoBehaviour.Expected_float : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Single Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Server_float_Function()' called when server was not active");
+                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Single Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Server_float_Function()' called when server was not active");
             }
             float actual = behaviour.Server_float_Function();
             Assert.AreEqual(expected, actual);
@@ -293,7 +144,7 @@ namespace Mirror.Tests.Attributes
             float expected = active ? AttributeBehaviour_MonoBehaviour.Expected_float : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Void Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Server_float_out_Function(System.Single&)' called when server was not active");
+                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Void Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Server_float_out_Function(System.Single&)' called when server was not active");
             }
             behaviour.Server_float_out_Function(out float actual);
             Assert.AreEqual(expected, actual);
@@ -308,7 +159,7 @@ namespace Mirror.Tests.Attributes
             ClassWithNoConstructor expected = active ? AttributeBehaviour_MonoBehaviour.Expected_ClassWithNoConstructor : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Server] function 'Mirror.Tests.Attributes.ClassWithNoConstructor Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Server_ClassWithNoConstructor_Function()' called when server was not active");
+                LogAssert.Expect(LogType.Warning, "[Server] function 'Mirror.Tests.EditorBehaviours.Attributes.ClassWithNoConstructor Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Server_ClassWithNoConstructor_Function()' called when server was not active");
             }
             ClassWithNoConstructor actual = behaviour.Server_ClassWithNoConstructor_Function();
             Assert.AreEqual(expected, actual);
@@ -323,7 +174,7 @@ namespace Mirror.Tests.Attributes
             ClassWithNoConstructor expected = active ? AttributeBehaviour_MonoBehaviour.Expected_ClassWithNoConstructor : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Void Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Server_ClassWithNoConstructor_out_Function(Mirror.Tests.Attributes.ClassWithNoConstructor&)' called when server was not active");
+                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Void Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Server_ClassWithNoConstructor_out_Function(Mirror.Tests.EditorBehaviours.Attributes.ClassWithNoConstructor&)' called when server was not active");
             }
             behaviour.Server_ClassWithNoConstructor_out_Function(out ClassWithNoConstructor actual);
             Assert.AreEqual(expected, actual);
@@ -338,7 +189,7 @@ namespace Mirror.Tests.Attributes
             ClassWithConstructor expected = active ? AttributeBehaviour_MonoBehaviour.Expected_ClassWithConstructor : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Server] function 'Mirror.Tests.Attributes.ClassWithConstructor Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Server_ClassWithConstructor_Function()' called when server was not active");
+                LogAssert.Expect(LogType.Warning, "[Server] function 'Mirror.Tests.EditorBehaviours.Attributes.ClassWithConstructor Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Server_ClassWithConstructor_Function()' called when server was not active");
             }
             ClassWithConstructor actual = behaviour.Server_ClassWithConstructor_Function();
             Assert.AreEqual(expected, actual);
@@ -353,7 +204,7 @@ namespace Mirror.Tests.Attributes
             ClassWithConstructor expected = active ? AttributeBehaviour_MonoBehaviour.Expected_ClassWithConstructor : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Void Mirror.Tests.Attributes.AttributeBehaviour_MonoBehaviour::Server_ClassWithConstructor_out_Function(Mirror.Tests.Attributes.ClassWithConstructor&)' called when server was not active");
+                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Void Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_MonoBehaviour::Server_ClassWithConstructor_out_Function(Mirror.Tests.EditorBehaviours.Attributes.ClassWithConstructor&)' called when server was not active");
             }
             behaviour.Server_ClassWithConstructor_out_Function(out ClassWithConstructor actual);
             Assert.AreEqual(expected, actual);
