+++ Assets/Mirror/Tests/Editor/Attributes/AttributeTests_NetworkBehaviour.cs
@@ -1,65 +1,13 @@
 // Tests for [Server], [Client], [ServerCallback], [ClientCallback] attributes.
 // Generated by AttributeTestGenerator.cs
 
+using Mirror.Tests.EditorBehaviours.Attributes;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
 
-namespace Mirror.Tests.Attributes
+namespace Mirror.Tests.EditorBehaviours.Attributes
 {
-    public class AttributeBehaviour_NetworkBehaviour : NetworkBehaviour
-    {
-        public static readonly float Expected_float = 2020f;
-
-        [Client]
-        public float Client_float_Function()
-        {
-            return Expected_float;
-        }
-
-        [Client]
-        public void Client_float_out_Function(out float value)
-        {
-            value = Expected_float;
-        }
-
-        [Server]
-        public float Server_float_Function()
-        {
-            return Expected_float;
-        }
-
-        [Server]
-        public void Server_float_out_Function(out float value)
-        {
-            value = Expected_float;
-        }
-
-        [ClientCallback]
-        public float ClientCallback_float_Function()
-        {
-            return Expected_float;
-        }
-
-        [ClientCallback]
-        public void ClientCallback_float_out_Function(out float value)
-        {
-            value = Expected_float;
-        }
-
-        [ServerCallback]
-        public float ServerCallback_float_Function()
-        {
-            return Expected_float;
-        }
-
-        [ServerCallback]
-        public void ServerCallback_float_out_Function(out float value)
-        {
-            value = Expected_float;
-        }
-    }
-
     public class AttributeTest_NetworkBehaviour
     {
         AttributeBehaviour_NetworkBehaviour behaviour;
@@ -91,7 +39,7 @@ namespace Mirror.Tests.Attributes
             float expected = active ? AttributeBehaviour_NetworkBehaviour.Expected_float : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Single Mirror.Tests.Attributes.AttributeBehaviour_NetworkBehaviour::Client_float_Function()' called when client was not active");
+                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Single Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_NetworkBehaviour::Client_float_Function()' called when client was not active");
             }
             float actual = behaviour.Client_float_Function();
             Assert.AreEqual(expected, actual);
@@ -106,7 +54,7 @@ namespace Mirror.Tests.Attributes
             float expected = active ? AttributeBehaviour_NetworkBehaviour.Expected_float : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Void Mirror.Tests.Attributes.AttributeBehaviour_NetworkBehaviour::Client_float_out_Function(System.Single&)' called when client was not active");
+                LogAssert.Expect(LogType.Warning, "[Client] function 'System.Void Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_NetworkBehaviour::Client_float_out_Function(System.Single&)' called when client was not active");
             }
             behaviour.Client_float_out_Function(out float actual);
             Assert.AreEqual(expected, actual);
@@ -121,7 +69,7 @@ namespace Mirror.Tests.Attributes
             float expected = active ? AttributeBehaviour_NetworkBehaviour.Expected_float : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Single Mirror.Tests.Attributes.AttributeBehaviour_NetworkBehaviour::Server_float_Function()' called when server was not active");
+                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Single Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_NetworkBehaviour::Server_float_Function()' called when server was not active");
             }
             float actual = behaviour.Server_float_Function();
             Assert.AreEqual(expected, actual);
@@ -136,7 +84,7 @@ namespace Mirror.Tests.Attributes
             float expected = active ? AttributeBehaviour_NetworkBehaviour.Expected_float : default;
             if (!active)
             {
-                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Void Mirror.Tests.Attributes.AttributeBehaviour_NetworkBehaviour::Server_float_out_Function(System.Single&)' called when server was not active");
+                LogAssert.Expect(LogType.Warning, "[Server] function 'System.Void Mirror.Tests.EditorBehaviours.Attributes.AttributeBehaviour_NetworkBehaviour::Server_float_out_Function(System.Single&)' called when server was not active");
             }
             behaviour.Server_float_out_Function(out float actual);
             Assert.AreEqual(expected, actual);
