+++ Assets/Mirror/Tests/EditorBehaviours/Mirror.Tests.EditorBehaviours.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 91ea2f53178f0594483565c76fd4f272
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
