+++ Assets/Mirror/Tests/EditorBehaviours/README.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1ae3def9d89797240b7c3ee01c3a6c5a
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
